chore: moved color folder to mobile/assets
